main backend/requirements.txt
fastapi==0.110.3
pandas
joblib==1.2.0
scikit-learn==1.3.2
numpy==1.23.5
uvicorn==0.23.2
Pyrebase4
tensorflow==2.16.1
pdfminer.six==20231228
spacy==2.3.9
